server/requirements.txt
mega.py==1.0.8
python-dotenv==0.19.0
